requirements.txt
fastapi
uvicorn[standard]
langchain
openai
pymilvus
psycopg2-binary
pydantic
python-dotenv